% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{check_response}
\alias{check_response}
\title{Check response for errors}
\usage{
check_response(response)
}
\arguments{
\item{response}{HTTP response}
}
\value{
string to determine retry strategy
}
\description{
If API returns 400, there was a problem with the provided parameters; stop
If API returns 400 and request was to /tokens endpoint, likely bad
credentials; stop.
If API returns 401, refresh access token if found and resend request.
If API returns 403 and request was to /tokens endpoint, sleep for 60 seconds
and try again.
If API returns 404, an invalid dataset name was provided; stop.
5xx errors are re-sent through a normal \code{RETRY}
}
