% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{count}
\alias{count}
\title{count}
\usage{
count(
  dataset,
  output = "df",
  manretry = 2,
  api_key = NULL,
  access_token = NULL,
  ...
)
}
\arguments{
\item{dataset}{A valid Enverus DirectAccess v2 API dataset as a string. For a
complete list of datasets, see \href{https://app.drillinginfo.com/direct/#/api/explorer/v2/gettingStarted}{Enverus API}}

\item{output}{A string indicating the type of output returned by the query:
\itemize{
\item \code{df} attempts to convert the query response into a \code{tibble} (the default).
\item \code{json} returns a raw list of lists
}}

\item{manretry}{Number of retries for user-related errors (authentication,
query parameters, etc.)}

\item{api_key}{A valid Enverus DirectAccess v2 API key as a string. If a key
is not provided, the query will attempt to use the \code{ENVERUS_API_KEY}
environment variable.}

\item{access_token}{A temporary token, as a string, authenticating the user.
If a token is not supplied, the query will attempt to authenticate the user
with the \code{ENVERUS_CLIENT_ID} and \code{ENVERUS_CLIENT_SECRET} environment
variables. See \code{\link[=get_access_token]{get_access_token()}}}

\item{...}{A named list of query filters, such as \code{deleteddate = "null"} to
be passed to the query. See the \href{https://app.drillinginfo.com/direct/#/api/explorer/v2/gettingStarted}{Enverus API}}
}
\description{
Count number of records in query
}
