% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{enverus2}
\alias{enverus2}
\title{enverus2}
\usage{
enverus2(dataset, ...)
}
\arguments{
\item{dataset}{A valid Enverus DirectAccess v2 API dataset as a string. For a
complete list of datasets, see
\url{https://app.drillinginfo.com/direct/#/api/explorer/v2/gettingStarted}}

\item{...}{A named list of query filters}
}
\description{
Query wrapper for the v2 API
}
